Object Oriented MaxMartenot
Added java objects for most of the major components of the program.
Biggest changes surround the serial input, which is now implemented as
an Iterator and the direct serial input is abstracted using the
ArduinoMessage object which is used as input for other objects. The
program's main method is significantly reduced in complexity using
chained function calls and constructor chaining. CODE IS NOT COMPLETE:
Still need to implement switch functionality in the Oscillator class,
but other than that, the other objects are complete.