Updated Object Oriented ProperMax
Added lots of code. Biggest changes since last push: SwitchArray object
has replaced the int array in the InstrumentData, allowing functions
that are passed an InstrumentData to check for a change in the state of
the switches without having to do the logic itself. Still leads to some
ugly code in the Oscillator class, but sadly there isn't a visible
parent for the oscillators we are using, so I can't create an array of
parent objects to hold our oscillators like I could for the Reverb
class. Implemented an AudioEffect stack, so we can add effects on top of
the audio signal, implemented as a stack of transfer functions on an
InstrumentData object. This simplifies lots of stuff, and opens the door
for code reuse in future instruments with different effects.